--- judo-sdk/src/main/java/com/judopay/UniqueIdentifier.java
@@ -1,11 +1,0 @@
-package com.judopay;
-
-import java.util.UUID;
-
-public class UniqueIdentifier {
-
-    public static String generate() {
-        return UUID.randomUUID().toString();
-    }
-
-}
