+++ judo-sdk/src/main/java/com/judopay/model/PaymentTransaction.java
@@ -1,6 +1,5 @@
 package com.judopay.model;
 
-import com.judopay.UniqueIdentifier;
 import com.judopay.api.Request;
 
 import java.util.Map;
