+++ judo-sdk/src/main/java/com/judopay/model/UniqueIdentifier.java
@@ -1,0 +1,11 @@
+package com.judopay.model;
+
+import java.util.UUID;
+
+class UniqueIdentifier {
+
+    static String generate() {
+        return UUID.randomUUID().toString();
+    }
+
+}
